[32X] fix SH-2 DMAC incorrectly being allowed to access CPU cache; this fixes Metal Head from crashing after the splash screen
